+++ packages/streams/src/MoralisStreams.ts
@@ -1,11 +1,11 @@
 import { readSettings, setSettings } from './resolvers';
-import { ApiConfig, Endpoints } from '@moralisweb3/api-utils';
-import { ApiModule, MoralisCore, MoralisCoreProvider, MoralisStreamError, StreamErrorCode } from '@moralisweb3/core';
+import { Endpoints } from '@moralisweb3/api-utils';
+import { ApiModule, MoralisCore, MoralisCoreProvider } from '@moralisweb3/core';
 import { createStream, CreateStreamOptions } from './methods/create';
 import { updateStream, UpdateStreamOptions } from './methods/update';
 import { deleteStream, DeleteStreamOptions } from './methods/delete';
 import { GetStreamsOptions, getStreams } from './methods/getAll';
-import { verifySignature, VerifySignatureOptions } from './utils/verifySignature';
+import { makeVerifySignature, VerifySignatureOptions } from './methods/utils';
 
 export const BASE_URL = 'https://streams-api.aws-prod-streams-master-1.moralis.io';
 
@@ -44,15 +44,5 @@ export class MoralisStreams extends ApiModule {
 
   public readonly readSettings = () => this._readSettings({});
 
-  public readonly verifySignature = (options: VerifySignatureOptions) => {
-    const apiKey = this.core.config.get(ApiConfig.apiKey);
-
-    if (!apiKey) {
-      throw new MoralisStreamError({
-        code: StreamErrorCode.GENERIC_STREAM_ERROR,
-        message: 'unable to verify signature without an api key',
-      });
-    }
-    return verifySignature(options, apiKey);
-  };
+  public readonly verifySignature = (options: VerifySignatureOptions) => makeVerifySignature(this.core)(options);
 }
